build: move common code into common webpack config
Move manifest build code into the common webpack config, because the
process for building the manifest is the exact same for dev and
prod, so this code should live in the common webpack config.
This is not the case currently as the same config is duplicated
in both dev and prod configs which means that if you change one, you
have to remember to change the other which is unideal.
