+++ webpack.dev.js
@@ -1,20 +1,8 @@
 const { merge } = require("webpack-merge");
 const common = require("./webpack.common.js");
-const webpack = require("webpack");
 const { sentryWebpackPlugin } = require("@sentry/webpack-plugin");
-const CopyPlugin = require("copy-webpack-plugin");
 
-const plugins = [
-  // Copies manifest.json into /build every build
-  new CopyPlugin({
-    patterns: [
-      { from: "./src/manifest.json", to: "./manifest.json" },
-      { from: "./src/ui" },
-      { from: "./src/assets", to: "./assets" },
-    ],
-  }),
-  new webpack.NormalModuleReplacementPlugin(/src[\\\/]env.ts/, "./env.dev.ts"),
-];
+const plugins = [];
 
 if (process.env.UPLOAD_SOURCEMAPS === "true") {
   plugins.push(
